Cleaned up warnings

+++ src/except.c
@@ -1,4 +1,6 @@
 #include <stdlib.h>
+#include <stdint.h>
+
 #include "exmemwb.h"
 
 // Clear pending exception after handled
@@ -24,21 +26,21 @@ void except_enter(const u32 exceptID)
     
     // Align the new SP to a frame
     cpu_set_sp((cpu_get_sp() - 0x20));// & ~0x4);
-    u32 * frame_ptr = (u32 *)cpu_get_sp();
+    u32 * frame_ptr = (u32 *)(uintptr_t)cpu_get_sp();
     
     // Exception can be mapped as a normal function call
     // so we need to backup the callee-saved registers for
     // the interrupted function
-    simStoreData((u32)&frame_ptr[0], cpu_get_gpr(0));
-    simStoreData((u32)&frame_ptr[1], cpu_get_gpr(1));
-    simStoreData((u32)&frame_ptr[2], cpu_get_gpr(2));
-    simStoreData((u32)&frame_ptr[3], cpu_get_gpr(3));
-    simStoreData((u32)&frame_ptr[4], cpu_get_gpr(12));
-    simStoreData((u32)&frame_ptr[5], cpu_get_lr());
-    simStoreData((u32)&frame_ptr[6], cpu_get_pc() - 0x4);
+    simStoreData((u32)(uintptr_t)&frame_ptr[0], cpu_get_gpr(0));
+    simStoreData((u32)(uintptr_t)&frame_ptr[1], cpu_get_gpr(1));
+    simStoreData((u32)(uintptr_t)&frame_ptr[2], cpu_get_gpr(2));
+    simStoreData((u32)(uintptr_t)&frame_ptr[3], cpu_get_gpr(3));
+    simStoreData((u32)(uintptr_t)&frame_ptr[4], cpu_get_gpr(12));
+    simStoreData((u32)(uintptr_t)&frame_ptr[5], cpu_get_lr());
+    simStoreData((u32)(uintptr_t)&frame_ptr[6], cpu_get_pc() - 0x4);
 
     u32 psr = cpu_get_apsr();
-    simStoreData((u32)&frame_ptr[7], (psr & 0xFFFFFC00) | (frame_align << 9) |  (psr & 0x1FF));
+    simStoreData((u32)(uintptr_t)&frame_ptr[7], (psr & 0xFFFFFC00) | (frame_align << 9) |  (psr & 0x1FF));
 
     // Encode the mode of the cpu at time of exception in LR value
     if(cpu_mode_is_handler())
@@ -89,31 +91,31 @@ void except_exit(const u32 pType)
     cpu_set_ipsr(0);
   
     // Restore registers
-    u32 * frame_ptr = (u32 *)cpu_get_sp();
+    u32 * frame_ptr = (u32 *)(uintptr_t)cpu_get_sp();
     u32 value;
     
-    simLoadData((u32)&frame_ptr[0], &value);
+    simLoadData((u32)(uintptr_t)&frame_ptr[0], &value);
     cpu_set_gpr(0, value);
     
-    simLoadData((u32)&frame_ptr[1], &value);
+    simLoadData((u32)(uintptr_t)&frame_ptr[1], &value);
     cpu_set_gpr(1, value);
     
-    simLoadData((u32)&frame_ptr[2], &value);
+    simLoadData((u32)(uintptr_t)&frame_ptr[2], &value);
     cpu_set_gpr(2, value);
     
-    simLoadData((u32)&frame_ptr[3], &value);
+    simLoadData((u32)(uintptr_t)&frame_ptr[3], &value);
     cpu_set_gpr(3, value);
     
-    simLoadData((u32)&frame_ptr[4], &value);
+    simLoadData((u32)(uintptr_t)&frame_ptr[4], &value);
     cpu_set_gpr(12, value);
     
-    simLoadData((u32)&frame_ptr[5], &value);
+    simLoadData((u32)(uintptr_t)&frame_ptr[5], &value);
     cpu_set_lr(value);
     
-    simLoadData((u32)&frame_ptr[6], &value);
+    simLoadData((u32)(uintptr_t)&frame_ptr[6], &value);
     cpu_set_pc(value);
     
-    simLoadData((u32)&frame_ptr[7], &value);
+    simLoadData((u32)(uintptr_t)&frame_ptr[7], &value);
     cpu_set_apsr(value);
     
     // Set special-purpose registers
